Fix build on wasm32-unknown-unknown
Rust only allows an empty match statement in code that will not be run,
that is when matching something that is not inhabited.

References to enums are always considered inhabited, because it may
avoid some undefined behaviour in unsafe code. See
https://github.com/rust-lang/rust/issues/78123

The solution in this PR is to match the enum rather than the reference
to the enum.

I also added CI for wasm32-unknown-unknown to combat regression.